fix: install spring bash

+++ install_spring.sh
@@ -2,34 +2,42 @@
 
 echo "########## INSTALLING SPRING BOOT ##########"
 
-echo "----- INSTALL OPENJDK 21 -----"
-sudo apt-get install openjdk-21-jdk
+echo "----- CHECKING OPENJDK 21 -----"
+if java -version 2>&1 | grep "openjdk version \"21" >/dev/null; then
+    echo "OpenJDK 21 is already installed."
+else
+    echo "OpenJDK 21 not found. Installing OpenJDK 21..."
+    sudo apt-get update
+    sudo apt-get install openjdk-21-jdk -y
+fi
 
-echo "----- MEASURE OPENJDK ALREADY INSTALLED -----"
-java -version
+echo "----- CHECKING MAVEN 3.9.9 -----"
+if mvn -version 2>&1 | grep "Apache Maven 3.9.9" >/dev/null; then
+    echo "Maven 3.9.9 is already installed."
+else
+    echo "Maven 3.9.9 not found. Downloading and installing Maven 3.9.9..."
+    wget https://dlcdn.apache.org/maven/maven-3/3.9.9/binaries/apache-maven-3.9.9-bin.tar.gz
 
-echo "----- DOWNLOAD MAVEN 3.9.9 -----"
-wget https://dlcdn.apache.org/maven/maven-3/3.9.9/binaries/apache-maven-3.9.9-bin.tar.gz
+    echo "----- EXTRACTING MAVEN 3.9.9 -----"
+    tar -xvf apache-maven-3.9.9-bin.tar.gz
+    rm apache-maven-3.9.9-bin.tar.gz
 
-echo "----- EXTRACTING MAVEN 3.9.9 -----"
-tar -xvf apache-maven-3.9.9-bin.tar.gz
-rm apache-maven-3.9.9-bin.tar.gz
+    echo "----- MOVING MAVEN TO /opt -----"
+    sudo mv apache-maven-3.9.9 /opt/
 
-echo "----- MOVE TO OPT -----"
-sudo mv apache-maven-3.9.9 /opt/
+    echo "----- SETTING UP ENV VARS FOR MAVEN -----"
+    VARIABLES='
+    export M2_HOME="/opt/apache-maven-3.9.9"
+    export PATH="$M2_HOME/bin:$PATH"
+    '
 
-echo "----- SETUP ENV VAR -----"
-VARIABLES='
-export M2_HOME="/opt/apache-maven-3.9.9"
-export PATH="$M2_HOME/bin:$PATH"
-'
+    echo "$VARIABLES" >> ~/.bashrc
 
-echo "$VARIABLES" >> ~/.bashrc
+    export M2_HOME="/opt/apache-maven-3.9.9"
+    export PATH="$M2_HOME/bin:$PATH"
+fi
 
-export M2_HOME="/opt/apache-maven-3.9.9"
-export PATH="$M2_HOME/bin:$PATH"
-
-echo "----- MEASURE MAVEN -----"
+echo "----- CHECKING MAVEN INSTALLATION -----"
 mvn -version
